refactor frontend to include presenter-container architecture
