+++ frontend/src/App.js
@@ -2,8 +2,8 @@ import './App.css';
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
 import {Toaster} from 'react-hot-toast';
 import Home from './pages/Home';
-import EditorPage from './pages/EditorPage';
 import {RecoilRoot} from "recoil";
+import EditorPageContainer from './pages/EditorPageContainer';
 
 function App() {
 
@@ -27,7 +27,7 @@ function App() {
                         <Route path="/" element={<Home />}></Route>
                         <Route
                             path="/editor/:roomId"
-                            element={<EditorPage />}
+                            element={<EditorPageContainer />}
                         ></Route>
                     </Routes>
                 </RecoilRoot>
